requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
drf-yasg==1.21.7
Faker==19.13.0
python-dotenv==1.0.0